Moved integration tests into tests/
