+++ tests/input/test_bai_error/allowed_instructions_a.txt
