+++ tests/input/test_bai_error/program.alpha
