+++ tests/input/test_cmd_check_compile_with_allowed_instructions_2/program.alpha
